i18n [nfc]: Add newly-known languages to the txln database, too.
The previous commit 204cda02a was meant to add to our list of
strings for translation the names of all the languages we have
translations for.  But before I merged it just now, I pushed
commit 3799aed94, which added a few new languages, and rebased
it on top of that.  So the commit ended up not adding *all* the
languages after all -- basically an undetected rebase conflict. :)

Fix the rebase, by rerunning the handy script provided in the
commit message.