no-messages [nfc]: Start untangling `Chat` UI selectors.
The selectors `getShowMessagePlaceholders` and `getIfNoMessages`
don't really belong as selectors: they're expressing UI choices,
and their logic belongs in UI code, to make the UI logic easier
to reason about.  The same is true of `canSendToNarrow`.

There's an obstacle to just inlining them, though, which is that
they have multiple call sites.  These are spread across the three
closely related components `Chat`, `MessageList`, and `NoMessages`,
and carry out tightly interdependent UI decisions -- so it's
important that these several call sites do use the same logic.

The more natural way to tie these together, more in keeping with
the architecture of a React app, is that the common ancestor of
these components should pass the information down as props.  Here,
that's `Chat`, and the other two are direct callees -- perfect for
that approach.

We'll do this in small NFC steps, to keep the changes simple to
reason about.  In this commit, we do some untangling of how these
functions are used together in the `Chat` parent component.