no-messages [nfc]: Pass `showMessagePlaceholders` down as a prop.
This is a UI choice, and one where it's important that these
neighboring components are all relying on the same logic for this
choice.  The proper React architectural pattern is therefore that
the common-ancestor component, here `Chat`, should pass the value
down as a prop.  Do that.

See preceding commit for further discussion.