no-messages [nfc]: Boil down logic for "say No Messages" condition.
This makes it more straightforward to answer the question:
when do we show the "No messages" message?

The pulling out of `isFetching` from `showMessagePlaceholders`
is NFC in the usual, almost purely synctactic, kind of way.
Then the edit to the definition of `sayNoMessages` is NFC too,
once you work through a bit of Boolean algebra.