store [nfc]: Fold setup of Redux middleware next to the store they go on.
The one thing this code does is construct a list of Redux middleware
values, to be passed to Redux when creating the Redux store.  These
values and their definitions only make sense in that context.  Move
them next to the code that provides that context.

Also add a bit of jsdoc.