store [nfc]: Give a less-terse name to imported library `redux-thunk`.
Just calling it "thunk" sounds like it's *a* thunk -- which is
either just wrong or all but vacuous, depending on how broad a
meaning one wants to give to the term "thunk".  Either way,
unhelpful.

A much more specific thing this is is that it's a Redux middleware --
just like `reactNavigationMiddleware`.  Name it similarly.