notifications [nfc]: Add new APNs-message validation function.
Hitherto we've been assuming that APNs messages matched a certain
format, and have been passing them upwards without validation.

In preparation to fix this, create a function that accepts a generic
object (as would be received from an APNs notification) and returns
either 1) a valid object of type Notification or 2) void.