eslint: Disable react-native/no-inline-styles.
Some styles are most readable when they are inlined.

This may be seen as a performance issue, because we're making new
style objects on each render, and sometimes this will result in an
unnecessary rerender. But the reality is we have tons of code that
does that already, because the lint rule doesn't catch even simple
variations. Some of these variations are very commonly used because
the style depends on props, state, and context -- so there's not a
clear path to turning this rule into something useful.

See discussion at
https://chat.zulip.org/#narrow/stream/243-mobile-team/topic/no-inline-styles.20rule/near/826995.