recipient [nfc]: Take message in getRecipientsIds, and check type.
Because the callers have been passing `message.display_recipient`,
whose type is `$FlowFixMe`, we don't have any type-checking of the
fact that they really are only passing PmRecipientUser arrays -- and
not stream names, as that property would be on a stream message.

So, just like pmKeyRecipientsFromMessage, take the whole `Message`
so the function can assert that fact just before it blithely starts
trying to interpret the `display_recipient` value.

This also requires patching up a couple of ill-typed test cases
to be a bit less so.