eventMiddleware: Reduce raciness of app-activity test.
Replace a check into the Redux store with a check of something closer
to the ground truth.

The Redux store's value may lag the actual state here based on any
number of factors -- and the actual state itself is one of them: while
inactive, we are likely to receive fewer and less frequent timeslices.

As a particular example: if a relevant message comes in just before
the user pushes the app into the background, we may end up without a
timeslice for some time. When we finally do process that message,
we'll end up trying to play the message-sound despite being in the
background, because we haven't yet been able to process the inactivity
event.

(N.B.: The summary correctly says "reduce", rather than "eliminate"!
We're still affected by whatever delays there are in the mechanism
that React Native itself uses to adjust the `AppState` object's value.

Even if we were to break down to native code ourselves, there would be
no guarantees available here: it will always be possible for the user
to suspend the app in the span between checking our activity status
and playing the message-sound.)