exampleData: Add some jsdoc on relationships between data.
It's often convenient in a test to have a message in a given stream,
a subscription to the same stream, etc.  Our baseline example objects
have the natural relationships of this kind, and deliberately so
because it is so often convenient and because it's easy to make new
unrelated objects when that's desired.

Document those relationships, to distinguish them from any accidental
arbitrary facts about these example objects, so that we can cheerfully
rely on them in tests.