api: Remove unused deleteStream for DELETE streams/${id}.
This one is not listed in #3540, but is in zulip/zulip#10044.

There, Tim says:

"I believe the app doesn't support editing these organization
settings. Further, it should be just getting these data from
/register with the right event_types value, so it's probably a
mobile bug if we're accessing these:"

"..."

"Same story with these stream settings/deactivation (though I could
imagine changing these):"

This endpoint is mentioned, but it's not for fetching data. Deleting
a stream is likely to be a nearer-term thing to add, but but this is
cheap to add later, when we end up using it, as Greg points out at
https://github.com/zulip/zulip-mobile/pull/3955#issuecomment-600885147.