api: Remove unused getAlertWords for GET users/me/alert_words.
For #3540, clean up code for unused endpoints. We don't expect to
use this one because the data is received on /register (see that
'alert_words' is present in `serverDataOnStartup` on the default
export of src/config.js).

This one was not specifically mentioned in #3540, but it was in
zulip/zulip#10044, which prompted it.

Fixes: #3540