docs: Add a "tips" file, to save various tidbits from code reviews.
Often when reviewing a PR I find that I write up for some point an
explanation that I feel might be helpful to someone else in the
future, in addition to the author of the PR at hand.

If it remains only in that PR thread on GitHub, though, it can be
quite challenging to pull up later when I want to share it with
someone else -- let alone for anyone to discover on their own when
they're looking for information on that subject.

Sometimes I go and add that information to our docs, and sometimes I
put it into a stream conversation on chat.zulip.org.  But there's also
a lot of these that accumulate, for which I haven't felt like we yet
have an appropriate place to put them.

So, here's an experiment, using one such PR comment that I wrote
yesterday.

Putting something into this file means not being very particular
about sorting it into a precisely appropriate place...
* but it at least makes it a lot easier to go back and find, if you
  remember having either written or read such a writeup, than it is
  as a comment in a GitHub PR thread.
* It might even have a chance at being discoverable, by someone
  looking for information having not previously seen the given
  writeup -- certainly it's unlikely to be *less* discoverable than
  a PR comment.
* And, by lowering the energy barrier compared to finding a more
  precisely appropriate place (or creating such a place, or doing
  more work to polish it for such a place), hopefully it will help
  with getting more of these collected at all that would otherwise
  languish in GitHub comment threads.