presence: Fix a potential race in reacting to app state change.
Update to the current state after subscribing to ongoing changes,
rather than before.

Otherwise, there's in principle a possible race where:
 * We call `updateHeartbeatState`, look at `AppState.currentState`,
   and it's one value.
 * Then the state changes, exogenously, to a new value.
 * Then we call `AppState.addEventListener` to start listening
   for changes.

In this scenario, we'd miss the update because we hadn't yet started
listening, and continue to think the state was its old value.

Because JavaScript is single-threaded, I think this can't actually
happen as long as `AppState.currentState` is just a property on a
plain JS object managed by pure JS code.  ... And looking at RN's
current implementation of `AppState` as I see it in my node_modules/,
it looks like that is the case, so this race can't currently happen.

But if `AppState.currentState` were something more exotic, it
potentially could, so it'd be nice not to have to study closely to
rule that out.  Because `heartbeat.toState` is idempotent, the
opposite order is guaranteed harmless; so do that.