docs/contributing: Cut CLA mention; refresh text on style.
We haven't required a CLA for a long time; see zulip/zulip@998b52dd8
and the previous background linked there.

This section of our contributor docs also had some unhelpful advice
(the link at "commit discipline" went to a page on code style choices
internal to the zulip/zulip codebase) and just hadn't been updated in
a while; so fix that and generally refresh it.

I haven't OTOH thought hard anew about what are the most important few
things to say to a new contributor, and whether this is the right list
of those.  So that's one way this file could likely still be improved.