webview build [nfc]: Flip sense of check-dest flag.
In a CLI, the defaults generally should favor interactive use,
because adding a flag is a much bigger burden on quick command-line
use than it is in a script.  These checks don't make sense when
using this script interactively, so that favors having them opt-in.

Moreover, the content of these particular checks is really that
they're checking up on whether the caller is doing the caller's own
job right -- and doing so on the presumption that the caller is our
actual build script for either iOS or Android.  (Nothing this tool
itself does depends on how the destination directory is named.
It'd be pretty weird if it did, and pretty questionable.)  The
checks are inappropriate in any other context, so that also favors
having them be opted into by the specific callers that need them.