tests [nfc]: Rename auxiliary/ to lib/.
This name is a more conventional one, for expressing relationships
like the one this code has to our tests: it's a set of libraries
(well, right now, a set of just one library) for writing tests.

It's also shorter. :-)  We'd originally chosen "aux", before
discovering that on Windows that filename is reserved due to some
poor design choices made in the 1970s:
  https://twitter.com/Foone/status/1058676834940776450
In 5b98965c0 (#3953), we renamed to "auxiliary" to fix that.
But "lib" is the same nice short length as "aux".