deps: Fix peer dependency warning with expo-application.
On expo-application v2.1.0 (see 18c37ce0f), we were seeing this
warning on initial runs of `yarn`:

warning " > expo-application@2.1.0" has unmet peer dependency "@unimodules/core@~1.0.0".

(Same as in 5b6014e83: "No such warning if `yarn` stops after step
'Resolving packages' and prints 'Already up-to-date'; but if it has
any actual work to do, it always shows this warning in step 'Linking
dependencies'.)

Version 2.1.0 of expo-application declares that it depends on a
version of @unimodules/core (1.0.0) that was already several months
out-of-date at the time expo-application was first created, and it
seems unlikely that this was well-considered.

I filed https://github.com/expo/expo/issues/7728 for this, and it
was promptly addressed by releasing version 2.1.1 of
expo-application that had "*" for its version constraint on the
@unimodules/core peer dependency.

So, switch to that new version.

Even when we do that, though, we get the following:

warning " > expo-application@2.1.1" has unmet peer dependency "@unimodules/core@*".

This goes away when we declare a version of @unimodules/core in our
own dependencies, though. Greg did some digging and concluded that
this is likely what we're supposed to do, actually, because it's
declared as a "peer dependency":

"""
Here's the reference doc for it:
https://docs.npmjs.com/files/package.json#peerdependencies but it's
not very clear about what it actually *means* -- mostly it just
describes a use-case.

The closest I've found to clear docs on it may be this Yarn blog
post:
https://classic.yarnpkg.com/blog/2018/04/18/dependencies-done-right/

> The `peerDependencies` object guarantees that, for each entry, any
> package that requires you will be requested to provide a copy of
> the dependency listed here, ideally matching the version you
> requested. It also guarantees that you’ll be able to access this
> dependency through `require()`, and finally **it also guarantees
> that the return of `require()` will be exactly the same version &
> instance than the one your parent would obtain if they were to
> `require()` it themselves**. [emphasis in original]

And if you work through the consequences of that... it means in
particular that the "parent" has to have that dependency and be able
to `require()` it themselves.

Which, if you're declaring your dependencies cleanly, means the
parent should have it as a dependency in their own `package.json`.
"""

So, declare @unimodules/core in our own "dependencies", with the
version specified by react-native-unimodules in its active version.
This matching should be done each time we change versions of
react-native-unimodules.