auth ui: Add iOS-compliant "Sign in with Apple" button component.
In Apple's Human Interface Guidelines doc about the button to be
used for "Sign in with Apple", they recommend using a standard
component from the iOS SDK if possible, but that you can use a
custom button, subject to design constraints, otherwise. That doc is
https://developer.apple.com/design/human-interface-guidelines/sign-in-with-apple/overview/buttons/.

So, make a component, only to be used on iOS, that will check if the
standard button is available, and use it, if so. The fallback custom
button will show on iOS <13.

Since Apple doesn't review our Google Play Store submissions, we
shouldn't even have to use the custom button on Android; we can just
use the same ZulipButton as the other social auth buttons use. I
think it looks better when they all look uniform like this.