api types [nfc]: Add minimal top-level comments on initial data types.
It'd be good to say more than this, but this is a start.  In
particular I spent a few seconds just now thinking the name
InitialDataUpdateMessageFlags, for what actually contains the
unread_msgs data structure, had to be a mistake, a copy-paste error,
before remembering how these type definitions are structured.